feat: add file-based configuration for serve command
Integrates the Viper package to allow the serve subcommand to be
configured via a file or the CLI. CLI takes precedence.

With this change the Viper instance becomes the authoritative source
for the configuration as it will manage both the Cobra CLI and the
configuration file. Unfortunately Viper does not support updating
configuration variables that are passed to Cobra by reference. This
means that all configuration state must be accessed via the Viper
API. Therefore this patch removes passing references to configuration
variables to Cobra and instead uses logic to initialize these
variables from Viper early in the command execution code.

Viper also does NOT honor any of Cobra's "required" meta data when
configuration data is read from a file (i.e. "MarkFlagRequired",
"MarkFlagMutuallyExclusive", etc. are ignored). Therefore the code
must manually check that required flags are set. This patch removes
the Cobra required flags settings and moves all validation checks to
where the configuration is pulled from the Viper API.

Only CLI and file-based configuration is supported - environment
variable-based configuration is TBD.

Closes: #44
Signed-off-by: Kenneth Giusti <kgiusti@redhat.com>